server migrate to apps
