+++ apps/servers/main/src/app.module.ts
@@ -1,0 +1,22 @@
+import { GreetingService } from '@application/services/greeting';
+import { IGreetingService } from '@application/services/interfaces/greeting';
+import { IGreetingRepository } from '@domain/interfaces/repositories/greeting';
+import { GreetingRepository } from '@infrastructure/repositories/greeting';
+import { Module } from '@nestjs/common';
+import { AppController } from '@presentation/controllers/app';
+
+@Module({
+  imports: [],
+  controllers: [AppController],
+  providers: [
+    {
+      provide: IGreetingService,
+      useClass: GreetingService,
+    },
+    {
+      provide: IGreetingRepository,
+      useClass: GreetingRepository,
+    },
+  ],
+})
+export class AppModule {}
