+++ package.json
@@ -4,72 +4,42 @@
   "private": true,
   "description": "",
   "license": "UNLICENSED",
-  "author": "",
+  "author": "TakashiAihara",
   "type": "module",
   "scripts": {
-    "analyze": "pnpm analyze:check",
-    "analyze:check": "biome check --verbose",
-    "analyze:check:ci": "pnpm analyze:check --log-level=debug --reporter=github",
-    "analyze:check:staged": "pnpm analyze:check --staged --no-errors-on-unmatched",
-    "analyze:check:unsafe": "pnpm analyze:check --unsafe",
-    "analyze:write": "biome check --write",
-    "build": "pnpm build:swc",
-    "build:swc": "nest --builder swc build",
-    "build:type": "tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json && fix-esm-import-path dist",
-    "format": "pnpm format:write && pnpm format:packages",
-    "format:check": "biome format",
-    "format:packages": "sort-package-json",
-    "format:write": "biome format --write",
-    "lint:check": "biome lint",
-    "lint:write": "biome lint --write",
-    "type:check": "tsc --noEmit",
-    "prepare": "husky",
-    "start": "nest --builder swc start",
-    "start:debug": "pnpm start:dev --debug --type-check",
-    "start:dev": "pnpm start --watch",
-    "start:prod": "node dist/main",
-    "test": "pnpm test:base",
-    "test:all": "pnpm test:unit && pnpm test:integration",
-    "test:unit": "pnpm test:base",
-    "test:base": "vitest",
-    "test:benchmark": "pnpm test:base --mode benchmark",
-    "test:cov": "pnpm test:base --coverage",
-    "test:debug": "pnpm test:base --inspect-brk --no-file-parallelism",
-    "test:debug:single-fork": "pnpm test:base --inspect-brk --pool forks --poolOptions.forks.singleFork",
-    "test:debug:single-thread": "pnpm test:base --inspect-brk --pool threads --poolOptions.threads.singleThread",
-    "test:integration": "pnpm test:base -c vitest.config.integration.ts",
-    "test:snapshot:update": "pnpm test:base --update --ui",
-    "test:watch": "pnpm test:base --watch"
-  },
-  "dependencies": {
-    "@nestjs/common": "^10.3.10",
-    "@nestjs/core": "^10.3.10",
-    "@nestjs/platform-express": "^10.3.10",
-    "reflect-metadata": "^0.2.2",
-    "rxjs": "^7.8.1"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "^1.8.3",
-    "@nestjs/cli": "^10.4.2",
-    "@nestjs/schematics": "^10.1.3",
-    "@nestjs/testing": "^10.3.10",
-    "@swc/cli": "^0.4.0",
-    "@swc/core": "^1.7.3",
-    "@tsconfig/strictest": "^2.0.5",
-    "@types/express": "^4.17.21",
-    "@types/node": "^22.0.0",
-    "@types/supertest": "^6.0.2",
-    "@vitest/coverage-v8": "^2.0.4",
-    "fix-esm-import-path": "^1.10.0",
-    "husky": "^9.1.4",
-    "sort-package-json": "^2.10.0",
-    "source-map-support": "^0.5.21",
-    "supertest": "^7.0.0",
-    "tsc-alias": "^1.8.10",
-    "tsx": "^4.16.3",
-    "typescript": "^5.5.4",
-    "unplugin-swc": "^1.5.1",
-    "vitest": "^2.0.4"
+    "analyze": "pnpm -r analyze",
+    "analyze:check": "pnpm -r analyze:check",
+    "analyze:check:ci": "pnpm -r analyze:check:ci",
+    "analyze:check:staged": "pnpm -r analyze:check:staged",
+    "analyze:check:unsafe": "pnpm -r analyze:check:unsafe",
+    "analyze:write": "pnpm -r analyze:write",
+    "build": "pnpm -r build",
+    "build:swc": "pnpm -r build:swc",
+    "build:type": "pnpm -r build:type",
+    "format": "pnpm -r format",
+    "format:check": "pnpm -r format:check",
+    "format:packages": "pnpm -r format:packages",
+    "format:write": "pnpm -r format:write",
+    "lint:check": "pnpm -r lint:check",
+    "lint:write": "pnpm -r lint:write",
+    "type:check": "pnpm -r type:check",
+    "prepare": "pnpm -r prepare",
+    "start": "pnpm -r start",
+    "start:debug": "pnpm -r start:debug",
+    "start:dev": "pnpm -r start:dev",
+    "start:prod": "pnpm -r start:prod",
+    "test": "pnpm -r test",
+    "test:all": "pnpm -r test:all",
+    "test:unit": "pnpm -r test:unit",
+    "test:base": "pnpm -r test:base",
+    "test:benchmark": "pnpm -r test:benchmark",
+    "test:cov": "pnpm -r test:cov",
+    "test:debug": "pnpm -r test:debug",
+    "test:debug:single-fork": "pnpm -r test:debug:single-fork",
+    "test:debug:single-thread": "pnpm -r test:debug:single-thread",
+    "test:integration": "pnpm -r test:integration",
+    "test:snapshot:update": "pnpm -r test:snapshot:update",
+    "test:watch": "pnpm -r test:watch"
   },
   "packageManager": "pnpm@9.6.0",
   "engines": {
