+++ pnpm-workspace.yaml
@@ -1,0 +1,4 @@
+packages:
+  - packages/*
+  - apps/**/*
+  - tools/*
